+++ slm_lab/agent/algorithm/__init__.py
@@ -7,6 +7,7 @@ Uses the nets module to build neural networks as the relevant function approxima
 # expose all the classes
 from .actor_critic import *
 from .dqn import *
+from .hydra_dqn import *
 from .ppo import *
 from .random import *
 from .reinforce import *
